pyproject.toml
[project]
name = "it-support-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Alonso",email = "alonavarro@estudiantec.cr"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "openai (>=1.90.0,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "chromadb (>=1.0.13,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "it_support_bot", from = "src"}]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
